refactor : chess controller view, rest controller 로 분리
